only delete existing indexes

+++ DBIx/Class/Elasticsearch/Role/ElasticSchema.pm
@@ -162,7 +162,7 @@ sub drop_indexes {
         if ( !$deleted_index->{ $rs->index_name } ) {
 
             $deleted_index->{ $rs->index_name } = 1;
-            $self->es->indices->delete( index => $rs->index_name );
+            $self->es->indices->delete( index => $rs->index_name ) if $self->es->indices->exists( index => $rs->index_name );
         }
     }
